+++ tests/unit/circuit-breaker.class.test.ts
@@ -66,7 +66,6 @@ describe("Circuit Breaker", () => {
     vi.useRealTimers();
     vi.restoreAllMocks();
   });
-
   describe("CircuitBreaker class", () => {
     describe("constructor and initialization", () => {
       it("should create circuit breaker with default options", async () => {
@@ -887,496 +886,3 @@ describe("Circuit Breaker", () => {
     });
   });
 
-  describe("Factory functions", () => {
-    describe("createCircuitBreaker", () => {
-      it("should create a circuit breaker with given name", () => {
-        const cb = createCircuitBreaker("my-cb");
-
-        expect(cb).toBeInstanceOf(CircuitBreaker);
-      });
-
-      it("should pass options to circuit breaker", async () => {
-        const cb = createCircuitBreaker("my-cb", {
-          failureThreshold: 3,
-          resetTimeoutMs: 10000,
-        });
-
-        // Test that options are used
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // 2 failures - still closed
-        for (let i = 0; i < 2; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-        expect(await cb.getState()).toBe("closed");
-
-        // 3rd failure - opens
-        try {
-          await cb.execute(fn);
-        } catch {
-          // expected
-        }
-        expect(await cb.getState()).toBe("open");
-      });
-    });
-
-    describe("createGitHubCircuitBreaker", () => {
-      it("should create circuit breaker for GitHub API", () => {
-        const cb = createGitHubCircuitBreaker(mockEnv);
-
-        expect(cb).toBeInstanceOf(CircuitBreaker);
-      });
-
-      it("should use GitHub-specific configuration", async () => {
-        const cb = createGitHubCircuitBreaker(mockEnv);
-
-        // Should use failureThreshold: 5
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // 4 failures - still closed
-        for (let i = 0; i < 4; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-        expect(await cb.getState()).toBe("closed");
-
-        // 5th failure - opens
-        try {
-          await cb.execute(fn);
-        } catch {
-          // expected
-        }
-        expect(await cb.getState()).toBe("open");
-      });
-
-      it("should use 30 second reset timeout", async () => {
-        const cb = createGitHubCircuitBreaker(mockEnv);
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // Open circuit
-        try {
-          await cb.execute(fn);
-        } catch {
-          // expected
-        }
-        for (let i = 0; i < 4; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-
-        vi.advanceTimersByTime(30000);
-
-        // Should transition to half-open
-        await cb.execute(vi.fn().mockResolvedValue("success"));
-        expect(await cb.getState()).toBe("half-open");
-      });
-    });
-
-    describe("createTelegramCircuitBreaker", () => {
-      it("should create circuit breaker for Telegram", () => {
-        const cb = createTelegramCircuitBreaker(mockEnv);
-
-        expect(cb).toBeInstanceOf(CircuitBreaker);
-      });
-
-      it("should use lower failure threshold for notifications", async () => {
-        const cb = createTelegramCircuitBreaker(mockEnv);
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // 2 failures - still closed
-        for (let i = 0; i < 2; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-        expect(await cb.getState()).toBe("closed");
-
-        // 3rd failure - opens (lower threshold)
-        try {
-          await cb.execute(fn);
-        } catch {
-          // expected
-        }
-        expect(await cb.getState()).toBe("open");
-      });
-
-      it("should use 60 second reset timeout", async () => {
-        const cb = createTelegramCircuitBreaker(mockEnv);
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // Open circuit
-        for (let i = 0; i < 3; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-
-        // Before 60 seconds - still open
-        vi.advanceTimersByTime(59999);
-        await expect(
-          cb.execute(vi.fn().mockResolvedValue("success")),
-        ).rejects.toThrow(CircuitBreakerOpenError);
-
-        // At 60 seconds - transitions
-        vi.advanceTimersByTime(1);
-        await cb.execute(vi.fn().mockResolvedValue("success"));
-        expect(await cb.getState()).toBe("half-open");
-      });
-    });
-
-    describe("getSourceCircuitBreaker", () => {
-      it("should create circuit breaker for domain", () => {
-        const cb = getSourceCircuitBreaker("example.com", mockEnv);
-
-        expect(cb).toBeInstanceOf(CircuitBreaker);
-      });
-
-      it("should return same instance for same domain", () => {
-        const cb1 = getSourceCircuitBreaker("example.com", mockEnv);
-        const cb2 = getSourceCircuitBreaker("example.com", mockEnv);
-
-        expect(cb1).toBe(cb2);
-      });
-
-      it("should return different instances for different domains", () => {
-        const cb1 = getSourceCircuitBreaker("example.com", mockEnv);
-        const cb2 = getSourceCircuitBreaker("other.com", mockEnv);
-
-        expect(cb1).not.toBe(cb2);
-      });
-
-      it("should use 5 minute reset timeout for sources", async () => {
-        const cb = getSourceCircuitBreaker("example.com", mockEnv);
-        const fn = vi.fn().mockRejectedValue(new Error("failure"));
-
-        // Open circuit
-        for (let i = 0; i < 5; i++) {
-          try {
-            await cb.execute(fn);
-          } catch {
-            // expected
-          }
-        }
-
-        // Before 5 minutes - still open
-        vi.advanceTimersByTime(299999);
-        await expect(
-          cb.execute(vi.fn().mockResolvedValue("success")),
-        ).rejects.toThrow(CircuitBreakerOpenError);
-
-        // At 5 minutes - transitions
-        vi.advanceTimersByTime(1);
-        await cb.execute(vi.fn().mockResolvedValue("success"));
-        expect(await cb.getState()).toBe("half-open");
-      });
-
-      it("should use source: prefix in name", async () => {
-        // Clear any existing circuit breaker
-        clearSourceCircuitBreakers();
-
-        // Create a fresh circuit breaker with env
-        const freshCb = getSourceCircuitBreaker("example.com", mockEnv);
-
-        // Open the circuit to force a state save
-        try {
-          await freshCb.execute(
-            vi.fn().mockRejectedValue(new Error("failure")),
-          );
-        } catch {
-          // expected
-        }
-        // Need 5 failures for source circuit breaker
-        for (let i = 0; i < 4; i++) {
-          try {
-            await freshCb.execute(
-              vi.fn().mockRejectedValue(new Error("failure")),
-            );
-          } catch {
-            // expected
-          }
-        }
-
-        expect(mockEnv.DEALS_PROD.put).toHaveBeenCalledWith(
-          "circuit:source:example.com",
-          expect.any(String),
-        );
-      });
-    });
-
-    describe("clearSourceCircuitBreakers", () => {
-      it("should clear all source circuit breakers", () => {
-        const cb1 = getSourceCircuitBreaker("example.com", mockEnv);
-        const cb2 = getSourceCircuitBreaker("other.com", mockEnv);
-
-        clearSourceCircuitBreakers();
-
-        const cb1Again = getSourceCircuitBreaker("example.com", mockEnv);
-        const cb2Again = getSourceCircuitBreaker("other.com", mockEnv);
-
-        expect(cb1).not.toBe(cb1Again);
-        expect(cb2).not.toBe(cb2Again);
-      });
-    });
-  });
-
-  describe("Convenience functions", () => {
-    describe("callWithCircuitBreaker", () => {
-      it("should execute function with circuit breaker", async () => {
-        const cb = new CircuitBreaker("test-cb", {}, mockEnv);
-        const fn = vi.fn().mockResolvedValue("result");
-
-        const result = await callWithCircuitBreaker(cb, fn);
-
-        expect(result).toBe("result");
-        expect(fn).toHaveBeenCalled();
-      });
-
-      it("should handle circuit breaker errors", async () => {
-        const cb = new CircuitBreaker(
-          "test-cb",
-          { failureThreshold: 1 },
-          mockEnv,
-        );
-
-        // Open circuit
-        try {
-          await cb.execute(vi.fn().mockRejectedValue(new Error("failure")));
-        } catch {
-          // expected
-        }
-
-        await expect(
-          callWithCircuitBreaker(cb, vi.fn().mockResolvedValue("success")),
-        ).rejects.toThrow(CircuitBreakerOpenError);
-      });
-    });
-
-    describe("getCircuitBreakerMetrics", () => {
-      it("should return metrics for circuit breaker", async () => {
-        const cb = new CircuitBreaker("test-cb", {}, mockEnv);
-        await cb.execute(vi.fn().mockResolvedValue("success"));
-
-        const metrics = getCircuitBreakerMetrics(cb);
-
-        expect(metrics.totalCalls).toBe(1);
-        expect(metrics.successfulCalls).toBe(1);
-      });
-    });
-
-    describe("resetCircuitBreaker", () => {
-      it("should reset circuit breaker to closed", async () => {
-        const cb = new CircuitBreaker(
-          "test-cb",
-          { failureThreshold: 1 },
-          mockEnv,
-        );
-
-        // Open circuit
-        try {
-          await cb.execute(vi.fn().mockRejectedValue(new Error("failure")));
-        } catch {
-          // expected
-        }
-
-        expect(await cb.getState()).toBe("open");
-
-        await resetCircuitBreaker(cb);
-
-        expect(await cb.getState()).toBe("closed");
-      });
-    });
-
-    describe("getAllCircuitBreakerMetrics", () => {
-      it("should return metrics for all circuit breakers", async () => {
-        const cb1 = new CircuitBreaker("cb1", {}, mockEnv);
-        const cb2 = new CircuitBreaker("cb2", {}, mockEnv);
-
-        await cb1.execute(vi.fn().mockResolvedValue("success"));
-
-        // Execute failing call and catch the error
-        try {
-          await cb2.execute(vi.fn().mockRejectedValue(new Error("failure")));
-        } catch {
-          // expected
-        }
-
-        const allMetrics = getAllCircuitBreakerMetrics();
-
-        expect(allMetrics["cb1"]).toBeDefined();
-        expect(allMetrics["cb2"]).toBeDefined();
-        expect(allMetrics["cb1"]!.successfulCalls).toBe(1);
-        expect(allMetrics["cb2"]!.failedCalls).toBe(1);
-      });
-
-      it("should return empty object when no metrics", () => {
-        const allMetrics = getAllCircuitBreakerMetrics();
-
-        expect(allMetrics).toEqual({});
-      });
-    });
-
-    describe("resetAllMetrics", () => {
-      it("should clear all metrics", async () => {
-        const cb = new CircuitBreaker("test-cb", {}, mockEnv);
-        await cb.execute(vi.fn().mockResolvedValue("success"));
-
-        expect(cb.getMetrics().totalCalls).toBe(1);
-
-        resetAllMetrics();
-
-        // After reset, metrics are re-initialized on next call
-        const cb2 = new CircuitBreaker("test-cb", {}, mockEnv);
-        expect(cb2.getMetrics().totalCalls).toBe(0);
-      });
-    });
-  });
-
-  describe("Full state transition cycle", () => {
-    it("should complete full cycle: closed → open → half-open → closed", async () => {
-      const cb = new CircuitBreaker(
-        "test-cb",
-        {
-          failureThreshold: 3,
-          resetTimeoutMs: 10000,
-          halfOpenMaxCalls: 2,
-        },
-        mockEnv,
-      );
-
-      const failFn = vi.fn().mockRejectedValue(new Error("failure"));
-      const successFn = vi.fn().mockResolvedValue("success");
-
-      // Start: CLOSED
-      expect(await cb.getState()).toBe("closed");
-
-      // 3 failures → OPEN
-      for (let i = 0; i < 3; i++) {
-        try {
-          await cb.execute(failFn);
-        } catch {
-          // expected
-        }
-      }
-      expect(await cb.getState()).toBe("open");
-
-      // Calls rejected in OPEN
-      await expect(cb.execute(successFn)).rejects.toThrow(
-        CircuitBreakerOpenError,
-      );
-
-      // After timeout → HALF-OPEN
-      vi.advanceTimersByTime(10000);
-      await cb.execute(successFn);
-      expect(await cb.getState()).toBe("half-open");
-
-      // Second success → CLOSED
-      await cb.execute(successFn);
-      expect(await cb.getState()).toBe("closed");
-
-      // Normal operation
-      await cb.execute(successFn);
-      await cb.execute(successFn);
-      expect(await cb.getState()).toBe("closed");
-
-      // Verify metrics
-      const metrics = cb.getMetrics();
-      expect(metrics.totalCalls).toBe(8); // 3 failures + 1 rejected + 2 to close + 2 normal
-      expect(metrics.successfulCalls).toBe(4);
-      expect(metrics.failedCalls).toBe(3);
-      expect(metrics.rejectedCalls).toBe(1);
-      expect(metrics.stateChanges).toBe(3); // closed→open, open→half-open, half-open→closed
-    });
-
-    it("should complete cycle with half-open failure: closed → open → half-open → open", async () => {
-      const cb = new CircuitBreaker(
-        "test-cb",
-        {
-          failureThreshold: 2,
-          resetTimeoutMs: 5000,
-          halfOpenMaxCalls: 3,
-        },
-        mockEnv,
-      );
-
-      const failFn = vi.fn().mockRejectedValue(new Error("failure"));
-      const successFn = vi.fn().mockResolvedValue("success");
-
-      // 2 failures → OPEN
-      for (let i = 0; i < 2; i++) {
-        try {
-          await cb.execute(failFn);
-        } catch {
-          // expected
-        }
-      }
-      expect(await cb.getState()).toBe("open");
-
-      // After timeout → HALF-OPEN
-      vi.advanceTimersByTime(5000);
-      await cb.execute(successFn);
-      expect(await cb.getState()).toBe("half-open");
-
-      // Failure in half-open → OPEN
-      try {
-        await cb.execute(failFn);
-      } catch {
-        // expected
-      }
-      expect(await cb.getState()).toBe("open");
-
-      // Verify state changes
-      const metrics = cb.getMetrics();
-      expect(metrics.stateChanges).toBe(3); // closed→open, open→half-open, half-open→open
-    });
-  });
-
-  describe("Multiple circuit breakers", () => {
-    it("should maintain separate state for different circuit breakers", async () => {
-      const cb1 = new CircuitBreaker("cb1", { failureThreshold: 2 }, mockEnv);
-      const cb2 = new CircuitBreaker("cb2", { failureThreshold: 5 }, mockEnv);
-
-      const failFn = vi.fn().mockRejectedValue(new Error("failure"));
-
-      // Open cb1 with 2 failures
-      for (let i = 0; i < 2; i++) {
-        try {
-          await cb1.execute(failFn);
-        } catch {
-          // expected
-        }
-      }
-
-      expect(await cb1.getState()).toBe("open");
-      expect(await cb2.getState()).toBe("closed");
-    });
-
-    it("should maintain separate metrics for different circuit breakers", async () => {
-      const cb1 = new CircuitBreaker("cb1", {}, mockEnv);
-      const cb2 = new CircuitBreaker("cb2", {}, mockEnv);
-
-      await cb1.execute(vi.fn().mockResolvedValue("success"));
-      await cb1.execute(vi.fn().mockResolvedValue("success"));
-      await cb2.execute(vi.fn().mockResolvedValue("success"));
-
-      const metrics1 = cb1.getMetrics();
-      const metrics2 = cb2.getMetrics();
-
-      expect(metrics1.totalCalls).toBe(2);
-      expect(metrics2.totalCalls).toBe(1);
-    });
-  });
-});
