--- tests/unit/d1-queries.test.ts
@@ -1,1265 +1,0 @@
-/**
- * Comprehensive Unit Tests for D1 Queries
- * Tests all query functions with mocked D1Database
- */
-
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import type {
-  D1Database,
-  D1PreparedStatement,
-  D1Result,
-} from "@cloudflare/workers-types";
-import type { Deal, ReferralInput } from "../../worker/types";
-import {
-  searchDeals,
-  getSearchSuggestions,
-  getDealsByDomain,
-  getDealsByCategory,
-  getDomainsWithCounts,
-  getCategoriesWithCounts,
-  getActiveDeals,
-  getExpiringDeals,
-  getRecentDeals,
-  getDealStats,
-  getDealTimeSeries,
-  insertDeal,
-  insertReferralCode,
-  getReferralCodesByDeal,
-  getReferralCodeByString,
-  getTopDomains,
-  getReferralUsageStats,
-  type DealSearchResult,
-  type DealStats,
-  type ExpiringDealRow,
-  type ReferralCodeResult,
-} from "../../worker/lib/d1/queries";
-
-// ============================================================================
-// Mock Factory
-// ============================================================================
-
-// Create the statement mock factory
-const createMockStatement = () => ({
-  bind: vi.fn().mockReturnThis(),
-  all: vi.fn().mockResolvedValue({ results: [], meta: {} }),
-  first: vi.fn().mockResolvedValue(null),
-  run: vi.fn().mockResolvedValue({ results: [], meta: {} }),
-});
-
-// Global statement reference that gets reset in beforeEach
-let currentMockStatement = createMockStatement();
-let currentMockSession: ReturnType<typeof createMockSession> | null = null;
-
-const createMockSession = () => {
-  const session = {
-    prepare: vi.fn().mockImplementation(() => currentMockStatement),
-    getBookmark: vi.fn().mockReturnValue("test-bookmark"),
-  };
-  return session;
-};
-
-const createMockD1 = () => {
-  // Reset the current statement and session
-  currentMockStatement = createMockStatement();
-  currentMockSession = createMockSession();
-
-  return {
-    prepare: vi.fn().mockImplementation(() => currentMockStatement),
-    batch: vi.fn().mockResolvedValue([]),
-    exec: vi.fn().mockResolvedValue(undefined),
-    withSession: vi.fn().mockImplementation(() => currentMockSession),
-  };
-};
-
-const createMockDeal = (
-  overrides: Partial<DealSearchResult> = {},
-): DealSearchResult => ({
-  id: 1,
-  deal_id: "deal-001",
-  title: "Test Deal",
-  description: "A test deal description",
-  domain: "example.com",
-  code: "TESTCODE",
-  url: "https://example.com/deal",
-  reward_type: "cash",
-  reward_value: 50,
-  reward_currency: "USD",
-  status: "active",
-  category: ["test", "demo"],
-  tags: ["new", "hot"],
-  confidence_score: 0.85,
-  ...overrides,
-});
-
-const createMockDealInput = (): Partial<Deal> & {
-  deal_id: string;
-  title: string;
-  url: string;
-  domain: string;
-} => ({
-  deal_id: "deal-002",
-  title: "New Test Deal",
-  description: "New deal description",
-  url: "https://newexample.com/deal",
-  domain: "newexample.com",
-  code: "NEWCODE",
-  source: {
-    url: "https://source.com",
-    domain: "source.com",
-    discovered_at: new Date().toISOString(),
-    trust_score: 0.8,
-  },
-  reward: {
-    type: "credit",
-    value: 100,
-    currency: "USD",
-    description: "$100 credit",
-  },
-  metadata: {
-    category: ["finance"],
-    tags: ["credit", "bonus"],
-    normalized_at: new Date().toISOString(),
-    confidence_score: 0.9,
-    status: "active",
-  },
-  expiry: {
-    date: new Date(Date.now() + 86400000).toISOString(),
-    confidence: 0.8,
-    type: "soft",
-  },
-  requirements: ["new user"],
-});
-
-const createMockReferralInput = (): ReferralInput & { deal_id: number } => ({
-  deal_id: 1,
-  url: "https://referral.com/code",
-  code: "REFCODE123",
-  domain: "referral.com",
-  description: "Referral code description",
-  source: "user_submitted",
-  status: "active",
-  submitted_at: new Date().toISOString(),
-  submitted_by: "user123",
-  metadata: {
-    title: "Referral Bonus",
-    reward_type: "cash",
-    reward_value: 25,
-    category: ["referral"],
-    tags: ["signup"],
-  },
-});
-
-// ============================================================================
-// Test Suite
-// ============================================================================
-
-describe("D1 Queries", () => {
-  let mockDb: ReturnType<typeof createMockD1>;
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-    mockDb = createMockD1();
-  });
-
-  // Helper to get current statement for assertions
-  const getMockStatement = () => currentMockStatement;
-
-  // Helper to get the session's prepare calls (for read operations using sessions)
-  const getSessionPrepareCalls = () => {
-    if (!currentMockSession) return [];
-    return currentMockSession.prepare.mock.calls;
-  };
-
-  // Helper to get last SQL query from session prepare (for read operations)
-  const getLastSessionQuery = () => {
-    const calls = getSessionPrepareCalls();
-    if (calls.length === 0) return null;
-    return calls[calls.length - 1]![0] as string;
-  };
-
-  // ============================================================================
-  // Full-Text Search Tests
-  // ============================================================================
-
-  describe("searchDeals", () => {
-    it("should return deals matching search query", async () => {
-      const mockDeals = [
-        createMockDeal(),
-        createMockDeal({ id: 2, deal_id: "deal-002" }),
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await searchDeals(
-        mockDb as unknown as D1Database,
-        "test",
-      );
-
-      expect(mockDb.withSession).toHaveBeenCalled();
-      expect(getMockStatement().bind).toHaveBeenCalled();
-      expect(results).toHaveLength(2);
-      expect(results[0]!.deal_id).toBe("deal-001");
-    });
-
-    it("should filter by status when provided", async () => {
-      const mockDeals = [createMockDeal({ status: "quarantined" })];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 1, rows_written: 0 },
-      });
-
-      const results = await searchDeals(
-        mockDb as unknown as D1Database,
-        "test",
-        {
-          status: "quarantined",
-        },
-      );
-
-      expect(results).toHaveLength(1);
-      expect(results[0]!.status).toBe("quarantined");
-    });
-
-    it("should include expired deals when includeExpired is true", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await searchDeals(mockDb as unknown as D1Database, "test", {
-        includeExpired: true,
-      });
-
-      const prepareCall = getLastSessionQuery();
-      expect(prepareCall).not.toContain("expiry_date IS NULL OR expiry_date >");
-    });
-
-    it("should respect limit option", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await searchDeals(mockDb as unknown as D1Database, "test", {
-        limit: 5,
-      });
-
-      const bindCalls = getMockStatement().bind.mock.calls;
-      const lastCall = bindCalls[bindCalls.length - 1];
-      expect(lastCall).toContain(5);
-    });
-
-    it("should return empty array on database error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await searchDeals(
-        mockDb as unknown as D1Database,
-        "test",
-      );
-
-      expect(results).toEqual([]);
-    });
-
-    it("should handle empty results", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      const results = await searchDeals(
-        mockDb as unknown as D1Database,
-        "nonexistent",
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getSearchSuggestions", () => {
-    it("should return title suggestions", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [{ title: "Deal One" }, { title: "Deal Two" }],
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const suggestions = await getSearchSuggestions(
-        mockDb as unknown as D1Database,
-        "dea",
-      );
-
-      expect(suggestions).toEqual(["Deal One", "Deal Two"]);
-    });
-
-    it("should filter out null titles", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [
-          { title: "Deal One" },
-          { title: null },
-          { title: "Deal Three" },
-        ],
-        success: true,
-        meta: { rows_read: 3, rows_written: 0 },
-      });
-
-      const suggestions = await getSearchSuggestions(
-        mockDb as unknown as D1Database,
-        "dea",
-      );
-
-      expect(suggestions).toEqual(["Deal One", "Deal Three"]);
-    });
-
-    it("should respect limit parameter", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getSearchSuggestions(mockDb as unknown as D1Database, "dea", 5);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith("dea", 5);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const suggestions = await getSearchSuggestions(
-        mockDb as unknown as D1Database,
-        "test",
-      );
-
-      expect(suggestions).toEqual([]);
-    });
-  });
-
-  // ============================================================================
-  // Domain and Category Tests
-  // ============================================================================
-
-  describe("getDealsByDomain", () => {
-    it("should return deals for specific domain", async () => {
-      const mockDeals = [
-        createMockDeal({ domain: "test.com" }),
-        createMockDeal({ id: 2, domain: "test.com", deal_id: "deal-002" }),
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getDealsByDomain(
-        mockDb as unknown as D1Database,
-        "test.com",
-      );
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(
-        "test.com",
-        expect.any(Number),
-      );
-      expect(results).toHaveLength(2);
-      expect(results[0]!.domain).toBe("test.com");
-    });
-
-    it("should include inactive deals when activeOnly is false", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getDealsByDomain(mockDb as unknown as D1Database, "test.com", {
-        activeOnly: false,
-      });
-
-      const prepareCall = getLastSessionQuery();
-      expect(prepareCall).not.toContain("is_active = 1 AND status = 'active'");
-    });
-
-    it("should respect limit option", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getDealsByDomain(mockDb as unknown as D1Database, "test.com", {
-        limit: 10,
-      });
-
-      const bindCalls = getMockStatement().bind.mock.calls;
-      const lastCall = bindCalls[bindCalls.length - 1];
-      expect(lastCall).toContain(10);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getDealsByDomain(
-        mockDb as unknown as D1Database,
-        "test.com",
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getDealsByCategory", () => {
-    it("should return deals in specific category", async () => {
-      const mockDeals = [createMockDeal({ category: ["finance", "crypto"] })];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 1, rows_written: 0 },
-      });
-
-      const results = await getDealsByCategory(
-        mockDb as unknown as D1Database,
-        "finance",
-      );
-
-      expect(results).toHaveLength(1);
-      expect(results[0]!.category).toContain("finance");
-    });
-
-    it("should filter by active status by default", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getDealsByCategory(mockDb as unknown as D1Database, "finance");
-
-      const bindCalls = getMockStatement().bind.mock.calls[0];
-      expect(bindCalls).toContain(1); // is_active = 1
-    });
-
-    it("should include inactive when activeOnly is false", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getDealsByCategory(mockDb as unknown as D1Database, "finance", {
-        activeOnly: false,
-      });
-
-      const bindCalls = getMockStatement().bind.mock.calls[0];
-      expect(bindCalls).toContain(0); // is_active = 0
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getDealsByCategory(
-        mockDb as unknown as D1Database,
-        "finance",
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getDomainsWithCounts", () => {
-    it("should return domains with deal counts", async () => {
-      const mockDomains = [
-        { domain: "example.com", count: 5 },
-        { domain: "test.com", count: 3 },
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDomains,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getDomainsWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toHaveLength(2);
-      expect(results[0]!.domain).toBe("example.com");
-      expect(results[0]!.count).toBe(5);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getDomainsWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getCategoriesWithCounts", () => {
-    it("should parse JSON categories and count them", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [
-          { categories: '["finance", "crypto"]' },
-          { categories: '["finance", "trading"]' },
-        ],
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getCategoriesWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toContainEqual({ name: "finance", count: 2 });
-      expect(results).toContainEqual({ name: "crypto", count: 1 });
-      expect(results).toContainEqual({ name: "trading", count: 1 });
-    });
-
-    it("should handle comma-separated categories as fallback", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [{ categories: "finance, crypto, trading" }],
-        success: true,
-        meta: { rows_read: 1, rows_written: 0 },
-      });
-
-      const results = await getCategoriesWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toContainEqual({ name: "finance", count: 1 });
-      expect(results).toContainEqual({ name: "crypto", count: 1 });
-      expect(results).toContainEqual({ name: "trading", count: 1 });
-    });
-
-    it("should sort by count descending", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [
-          { categories: '["popular"]' },
-          { categories: '["popular", "rare"]' },
-          { categories: '["popular"]' },
-        ],
-        success: true,
-        meta: { rows_read: 3, rows_written: 0 },
-      });
-
-      const results = await getCategoriesWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results[0]!.name).toBe("popular");
-      expect(results[0]!.count).toBe(3);
-    });
-
-    it("should return empty array on database error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getCategoriesWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toEqual([]);
-    });
-
-    it("should return empty array when no categories found", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      const results = await getCategoriesWithCounts(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  // ============================================================================
-  // Status-Based Query Tests
-  // ============================================================================
-
-  describe("getActiveDeals", () => {
-    it("should return only active deals", async () => {
-      const mockDeals = [
-        createMockDeal(),
-        createMockDeal({ id: 2, deal_id: "deal-002" }),
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getActiveDeals(mockDb as unknown as D1Database);
-
-      expect(results).toHaveLength(2);
-      expect(getMockStatement().bind).toHaveBeenCalledWith(expect.any(Number));
-    });
-
-    it("should respect limit parameter", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getActiveDeals(mockDb as unknown as D1Database, 25);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(25);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getActiveDeals(mockDb as unknown as D1Database);
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getExpiringDeals", () => {
-    it("should return deals expiring within specified days", async () => {
-      const mockDeals: ExpiringDealRow[] = [
-        {
-          id: 1,
-          deal_id: "deal-001",
-          title: "Expiring Soon",
-          domain: "test.com",
-          expiry_date: new Date(Date.now() + 86400000).toISOString(),
-          days_remaining: 1,
-          code: "EXPIRING",
-        },
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 1, rows_written: 0 },
-      });
-
-      const results = await getExpiringDeals(
-        mockDb as unknown as D1Database,
-        7,
-      );
-
-      expect(results).toHaveLength(1);
-      expect(results[0]!.days_remaining).toBe(1);
-    });
-
-    it("should default to 7 days", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getExpiringDeals(mockDb as unknown as D1Database);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(7);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getExpiringDeals(mockDb as unknown as D1Database);
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getRecentDeals", () => {
-    it("should return deals added within specified days", async () => {
-      const mockDeals = [
-        createMockDeal(),
-        createMockDeal({ id: 2, deal_id: "deal-002" }),
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDeals,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getRecentDeals(
-        mockDb as unknown as D1Database,
-        7,
-        50,
-      );
-
-      expect(results).toHaveLength(2);
-      expect(getMockStatement().bind).toHaveBeenCalledWith(7, 50);
-    });
-
-    it("should use defaults when not specified", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getRecentDeals(mockDb as unknown as D1Database);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(7, 50);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getRecentDeals(mockDb as unknown as D1Database);
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  // ============================================================================
-  // Statistics Query Tests
-  // ============================================================================
-
-  describe("getDealStats", () => {
-    it("should return comprehensive deal statistics", async () => {
-      // Setup mock responses for sequential queries in getDealStats
-      getMockStatement()
-        .run.mockResolvedValueOnce({
-          results: [
-            {
-              total: 100,
-              active: 80,
-              quarantined: 10,
-              rejected: 5,
-              expired: 5,
-            },
-          ],
-          success: true,
-          meta: { rows_read: 1, rows_written: 0 },
-        })
-        .mockResolvedValueOnce({
-          results: [
-            { domain: "example.com", count: 20 },
-            { domain: "test.com", count: 15 },
-          ],
-          success: true,
-          meta: { rows_read: 2, rows_written: 0 },
-        })
-        .mockResolvedValueOnce({
-          results: [
-            { type: "cash", count: 40 },
-            { type: "credit", count: 30 },
-          ],
-          success: true,
-          meta: { rows_read: 2, rows_written: 0 },
-        })
-        .mockResolvedValueOnce({
-          results: [{ categories: '["finance"]' }],
-          success: true,
-          meta: { rows_read: 1, rows_written: 0 },
-        });
-
-      const stats = await getDealStats(mockDb as unknown as D1Database);
-
-      expect(stats.total).toBe(100);
-      expect(stats.active).toBe(80);
-      expect(stats.quarantined).toBe(10);
-      expect(stats.rejected).toBe(5);
-      expect(stats.expired).toBe(5);
-      expect(stats.byDomain).toHaveLength(2);
-      expect(stats.byRewardType).toHaveLength(2);
-      expect(stats.byCategory).toHaveLength(1);
-    });
-
-    it("should handle zeros when no data", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      const stats = await getDealStats(mockDb as unknown as D1Database);
-
-      expect(stats.total).toBe(0);
-      expect(stats.active).toBe(0);
-      expect(stats.byDomain).toEqual([]);
-      expect(stats.byCategory).toEqual([]);
-      expect(stats.byRewardType).toEqual([]);
-    });
-
-    it("should return defaults on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const stats = await getDealStats(mockDb as unknown as D1Database);
-
-      expect(stats.total).toBe(0);
-      expect(stats.active).toBe(0);
-      expect(stats.byDomain).toEqual([]);
-    });
-  });
-
-  describe("getDealTimeSeries", () => {
-    it("should return time-series data for specified days", async () => {
-      const mockData = [
-        { date: "2024-01-01", count: 10, new_count: 2 },
-        { date: "2024-01-02", count: 12, new_count: 2 },
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockData,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getDealTimeSeries(
-        mockDb as unknown as D1Database,
-        30,
-      );
-
-      expect(results).toHaveLength(2);
-      expect(results[0]!.date).toBe("2024-01-01");
-      expect(results[0]!.count).toBe(10);
-    });
-
-    it("should default to 30 days", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getDealTimeSeries(mockDb as unknown as D1Database);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(30);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getDealTimeSeries(mockDb as unknown as D1Database);
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  // ============================================================================
-  // Insert/Update Tests
-  // ============================================================================
-
-  describe("insertDeal", () => {
-    it("should insert a new deal successfully", async () => {
-      getMockStatement().run.mockResolvedValue({
-        success: true,
-        meta: { last_row_id: 123, changes: 1 },
-      });
-
-      const deal = createMockDealInput();
-      const result = await insertDeal(mockDb as unknown as D1Database, deal);
-
-      expect(result.success).toBe(true);
-      expect(result.id).toBe(123);
-    });
-
-    it("should handle missing optional fields gracefully", async () => {
-      getMockStatement().run.mockResolvedValue({
-        success: true,
-        meta: { last_row_id: 456, changes: 1 },
-      });
-
-      const minimalDeal = {
-        deal_id: "deal-minimal",
-        title: "Minimal Deal",
-        url: "https://minimal.com",
-        domain: "minimal.com",
-      };
-
-      const result = await insertDeal(
-        mockDb as unknown as D1Database,
-        minimalDeal,
-      );
-
-      expect(result.success).toBe(true);
-    });
-
-    it("should return error on insert failure", async () => {
-      // D1Client returns error when run() throws
-      getMockStatement().run.mockRejectedValue(new Error("Duplicate deal_id"));
-
-      const deal = createMockDealInput();
-      const result = await insertDeal(mockDb as unknown as D1Database, deal);
-
-      expect(result.success).toBe(false);
-      expect(result.error).toContain("Duplicate deal_id");
-    });
-
-    it("should handle database errors", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Connection lost"));
-
-      const deal = createMockDealInput();
-      const result = await insertDeal(mockDb as unknown as D1Database, deal);
-
-      expect(result.success).toBe(false);
-      expect(result.error).toBeDefined();
-    });
-
-    it("should use default values for missing reward data", async () => {
-      getMockStatement().run.mockResolvedValue({
-        success: true,
-        meta: { last_row_id: 789, changes: 1 },
-      });
-
-      const dealWithoutReward = {
-        deal_id: "deal-no-reward",
-        title: "No Reward Deal",
-        url: "https://noreward.com",
-        domain: "noreward.com",
-      };
-
-      const result = await insertDeal(
-        mockDb as unknown as D1Database,
-        dealWithoutReward,
-      );
-
-      expect(result.success).toBe(true);
-    });
-  });
-
-  describe("insertReferralCode", () => {
-    it("should insert a referral code successfully", async () => {
-      getMockStatement().run.mockResolvedValue({
-        success: true,
-        meta: { last_row_id: 100, changes: 1 },
-      });
-
-      const referral = createMockReferralInput();
-      const result = await insertReferralCode(
-        mockDb as unknown as D1Database,
-        referral,
-      );
-
-      expect(result.success).toBe(true);
-      expect(result.id).toBe(100);
-    });
-
-    it("should handle referral without metadata", async () => {
-      getMockStatement().run.mockResolvedValue({
-        success: true,
-        meta: { last_row_id: 101, changes: 1 },
-      });
-
-      const minimalReferral = {
-        deal_id: 1,
-        url: "https://test.com",
-        code: "MINIMAL",
-      };
-
-      const result = await insertReferralCode(
-        mockDb as unknown as D1Database,
-        minimalReferral as ReferralInput & { deal_id: number },
-      );
-
-      expect(result.success).toBe(true);
-    });
-
-    it("should return error on insert failure", async () => {
-      // D1Client returns error when run() throws
-      getMockStatement().run.mockRejectedValue(
-        new Error("Code already exists"),
-      );
-
-      const referral = createMockReferralInput();
-      const result = await insertReferralCode(
-        mockDb as unknown as D1Database,
-        referral,
-      );
-
-      expect(result.success).toBe(false);
-      expect(result.error).toContain("Code already exists");
-    });
-
-    it("should handle database errors", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database timeout"));
-
-      const referral = createMockReferralInput();
-      const result = await insertReferralCode(
-        mockDb as unknown as D1Database,
-        referral,
-      );
-
-      expect(result.success).toBe(false);
-      expect(result.error).toBeDefined();
-    });
-  });
-
-  // ============================================================================
-  // Referral Code Query Tests
-  // ============================================================================
-
-  describe("getReferralCodesByDeal", () => {
-    it("should return referral codes for a deal", async () => {
-      const mockReferrals: ReferralCodeResult[] = [
-        {
-          id: 1,
-          code: "REF001",
-          deal_id: 1,
-          deal_title: "Test Deal",
-          domain: "test.com",
-          status: "active",
-          max_uses: 100,
-          current_uses: 50,
-          use_count: 50,
-        },
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockReferrals,
-        success: true,
-        meta: { rows_read: 1, rows_written: 0 },
-      });
-
-      const results = await getReferralCodesByDeal(
-        mockDb as unknown as D1Database,
-        1,
-      );
-
-      expect(results).toHaveLength(1);
-      expect(results[0]!.code).toBe("REF001");
-      expect(getMockStatement().bind).toHaveBeenCalledWith(1);
-    });
-
-    it("should filter by active status by default", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getReferralCodesByDeal(mockDb as unknown as D1Database, 1);
-
-      const prepareCall = getLastSessionQuery();
-      expect(prepareCall).toContain(
-        "rc.is_active = 1 AND rc.status = 'active'",
-      );
-    });
-
-    it("should include inactive when activeOnly is false", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getReferralCodesByDeal(mockDb as unknown as D1Database, 1, false);
-
-      const prepareCall = getLastSessionQuery();
-      expect(prepareCall).not.toContain(
-        "rc.is_active = 1 AND rc.status = 'active'",
-      );
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getReferralCodesByDeal(
-        mockDb as unknown as D1Database,
-        1,
-      );
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getReferralCodeByString", () => {
-    it("should return referral code by code string", async () => {
-      const mockReferral: ReferralCodeResult = {
-        id: 1,
-        code: "SPECIALCODE",
-        deal_id: 1,
-        deal_title: "Special Deal",
-        domain: "special.com",
-        status: "active",
-        max_uses: 100,
-        current_uses: 0,
-        use_count: 0,
-        expires_at: undefined,
-        days_remaining: undefined,
-      };
-      getMockStatement().first.mockResolvedValue(mockReferral);
-
-      const result = await getReferralCodeByString(
-        mockDb as unknown as D1Database,
-        "SPECIALCODE",
-      );
-
-      expect(result).not.toBeNull();
-      expect(result?.code).toBe("SPECIALCODE");
-      expect(getMockStatement().bind).toHaveBeenCalledWith("SPECIALCODE");
-    });
-
-    it("should be case-insensitive", async () => {
-      getMockStatement().first.mockResolvedValue(null);
-
-      await getReferralCodeByString(
-        mockDb as unknown as D1Database,
-        "lowercase",
-      );
-
-      const prepareCall = getLastSessionQuery();
-      expect(prepareCall).toContain("COLLATE NOCASE");
-    });
-
-    it("should return null when code not found", async () => {
-      getMockStatement().first.mockResolvedValue(null);
-
-      const result = await getReferralCodeByString(
-        mockDb as unknown as D1Database,
-        "NONEXISTENT",
-      );
-
-      expect(result).toBeNull();
-    });
-
-    it("should return null on error", async () => {
-      getMockStatement().first.mockRejectedValue(new Error("Database error"));
-
-      const result = await getReferralCodeByString(
-        mockDb as unknown as D1Database,
-        "TEST",
-      );
-
-      expect(result).toBeNull();
-    });
-  });
-
-  // ============================================================================
-  // Analytics Query Tests
-  // ============================================================================
-
-  describe("getTopDomains", () => {
-    it("should return top domains with deal and referral counts", async () => {
-      const mockDomains = [
-        { domain: "example.com", deals: 10, referrals: 25 },
-        { domain: "test.com", deals: 8, referrals: 15 },
-      ];
-      getMockStatement().run.mockResolvedValue({
-        results: mockDomains,
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const results = await getTopDomains(mockDb as unknown as D1Database, 10);
-
-      expect(results).toHaveLength(2);
-      expect(results[0]!.domain).toBe("example.com");
-      expect(results[0]!.deals).toBe(10);
-      expect(results[0]!.referrals).toBe(25);
-    });
-
-    it("should respect limit parameter", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getTopDomains(mockDb as unknown as D1Database, 5);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(5);
-    });
-
-    it("should default to 10 domains", async () => {
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getTopDomains(mockDb as unknown as D1Database);
-
-      expect(getMockStatement().bind).toHaveBeenCalledWith(10);
-    });
-
-    it("should return empty array on error", async () => {
-      getMockStatement().run.mockRejectedValue(new Error("Database error"));
-
-      const results = await getTopDomains(mockDb as unknown as D1Database);
-
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("getReferralUsageStats", () => {
-    it("should return comprehensive usage statistics", async () => {
-      getMockStatement()
-        .first.mockResolvedValueOnce({ count: 150 }) // totalUses
-        .mockResolvedValueOnce({ count: 45 }); // uniqueUsers
-
-      getMockStatement().run.mockResolvedValue({
-        results: [
-          { date: "2024-01-01", count: 5 },
-          { date: "2024-01-02", count: 10 },
-        ],
-        success: true,
-        meta: { rows_read: 2, rows_written: 0 },
-      });
-
-      const stats = await getReferralUsageStats(
-        mockDb as unknown as D1Database,
-        30,
-      );
-
-      expect(stats.totalUses).toBe(150);
-      expect(stats.uniqueUsers).toBe(45);
-      expect(stats.byDay).toHaveLength(2);
-    });
-
-    it("should handle missing data gracefully", async () => {
-      getMockStatement()
-        .first.mockResolvedValueOnce(null) // totalUses
-        .mockResolvedValueOnce(null); // uniqueUsers
-
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      const stats = await getReferralUsageStats(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(stats.totalUses).toBe(0);
-      expect(stats.uniqueUsers).toBe(0);
-      expect(stats.byDay).toEqual([]);
-    });
-
-    it("should default to 30 days", async () => {
-      getMockStatement()
-        .first.mockResolvedValueOnce({ count: 0 })
-        .mockResolvedValueOnce({ count: 0 });
-
-      getMockStatement().run.mockResolvedValue({
-        results: [],
-        success: true,
-        meta: { rows_read: 0, rows_written: 0 },
-      });
-
-      await getReferralUsageStats(mockDb as unknown as D1Database);
-
-      const firstCall = getMockStatement().bind.mock.calls[0];
-      expect(firstCall).toContain(30);
-    });
-
-    it("should return zeros on error", async () => {
-      getMockStatement().first.mockRejectedValue(new Error("Database error"));
-
-      const stats = await getReferralUsageStats(
-        mockDb as unknown as D1Database,
-      );
-
-      expect(stats.totalUses).toBe(0);
-      expect(stats.uniqueUsers).toBe(0);
-      expect(stats.byDay).toEqual([]);
-    });
-  });
-});
