+++ tests/unit/validation.batch-operations.test.ts
@@ -1,0 +1,102 @@
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import type { Deal, Env } from "../../worker/types";
+import { setGitHubToken } from "../../worker/lib/github/index";
+
+// Mock implementations for validation functions
+const mockValidateUrl = vi.fn();
+const mockCheckUrlStatusBatch = vi.fn();
+const mockDetectRedirects = vi.fn();
+const mockValidateCodeFormat = vi.fn();
+const mockValidateCodeOnPage = vi.fn();
+const mockTestCodeRedemption = vi.fn();
+const mockScrapeCurrentRewards = vi.fn();
+const mockDetectRewardChanges = vi.fn();
+const mockExtractRewardFromHTML = vi.fn();
+
+// ============================================================================
+// Test Fixtures
+// ============================================================================
+
+const createMockDeal = (
+  id: string,
+  overrides: Partial<Deal> & { expiryDate?: string } = {},
+): Deal => {
+  const expiryDate =
+    overrides.expiryDate ||
+    new Date(Date.now() + 30 * 24 * 60 * 60 * 1000).toISOString();
+
+  return {
+    id,
+    source: {
+      url: "https://example.com/invite",
+      domain: overrides.source?.domain || "example.com",
+      discovered_at: "2024-03-31T00:00:00Z",
+      trust_score: overrides.source?.trust_score || 0.7,
+    },
+    title: overrides.title ?? "Test Deal",
+    description: overrides.description ?? "Test description",
+    code: overrides.code ?? "CODE123",
+    url: overrides.url ?? "https://example.com/invite/CODE123",
+    reward: overrides.reward ?? {
+      type: "cash",
+      value: 50,
+      currency: "USD",
+    },
+    expiry: {
+      date: expiryDate,
+      confidence: overrides.expiry?.confidence ?? 0.8,
+      type: overrides.expiry?.type ?? "soft",
+    },
+    metadata: {
+      category: ["test"],
+      tags: ["test"],
+      normalized_at: "2024-03-31T00:00:00Z",
+      confidence_score: 0.8,
+      status:
+        (overrides.metadata?.status as "active" | "quarantined" | "rejected") ??
+        "active",
+    },
+  };
+};
+
+// ============================================================================
+// URL Validator Tests
+// ============================================================================
+
+describe("Batch Operations", () => {
+  it("should process URLs in batches with rate limiting", async () => {
+    const urls = [
+      "https://example1.com",
+      "https://example2.com",
+      "https://example3.com",
+    ];
+
+    const results = await Promise.all(
+      urls.map(async (url, i) => ({
+        url,
+        valid: true,
+        responseTimeMs: 100 + i * 50,
+        timestamp: new Date().toISOString(),
+      })),
+    );
+
+    expect(results).toHaveLength(3);
+    expect(results.every((r) => r.valid)).toBe(true);
+  });
+
+  it("should handle partial failures in batch", async () => {
+    const results = [
+      { url: "https://ok.com", valid: true },
+      { url: "https://error.com", valid: false, error: "Timeout" },
+      { url: "https://ok2.com", valid: true },
+    ];
+
+    expect(results.filter((r) => r.valid).length).toBe(2);
+    expect(results.filter((r) => !r.valid).length).toBe(1);
+  });
+});
+
+// ============================================================================
+// Integration Tests
+// ============================================================================
+
